pyproject.toml
[tool.poetry]
name = "wafflebot"
version = "0.1.0"
description = "A general purpose Discord bot"
authors = ["Alex Krantz <alex@krantz.dev>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aioredis = "^1.3.1"
Authlib = "^0.15.3"
httpx = "^0.17.0"
fastapi = "^0.68.0"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.22"}
uvicorn = {version = "^0.13.4", extras = ["standard"]}
ujson = "^4.0.2"
websockets = "^8.1"
pydantic = "^1.8.1"
"discord.py" = { git = "https://github.com/Rapptz/discord.py", rev = "58ca9e99" }
psycopg2-binary = "^2.8.6"
alembic = "^1.5.6"
python-dotenv = "^0.15.0"
cbor2 = "^5.2.0"
aiofiles = "^0.6.0"
itsdangerous = "^1.1.0"
asyncpg = "^0.22.0"
dateparser = "^1.0.0"
gunicorn = "^20.1.0"
sentry-sdk = "^1.3.1"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
ipython = "^7.21.0"
black = "^20.8b1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
